feat: embedded entity Image 생성 #30
dish에 필요한 image들을 모아서 객체로 관리하기 위해 embedded entity 생성
