+++ BE/src/main/java/com/codesquad/sidedish/domain/Dish.java
@@ -2,12 +2,13 @@ package com.codesquad.sidedish.domain;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Column;
+import org.springframework.data.relational.core.mapping.Embedded;
 
 public class Dish {
 
     @Id
     private String id;
-    private String mainImage;
+
     private String alt;
     private String deliveryType;
     private String title;
@@ -16,22 +17,21 @@ public class Dish {
     private String sellingPrice;
     private String badge;
     private Long categoryId;
-    private String topImage;
-    private String thumbImages;
     private String deliveryInfo;
     private String deliveryFee;
-    private String detailSection;
     private int stock;
     private String point;
+
+    @Embedded.Nullable
+    private Image image;
 
     private Dish() {
     }
 
-    public Dish(String id, String mainImage, String alt, String deliveryType, String title, String description,
-                String normalPrice, String sellingPrice, String badge, Long categoryId, String topImage, String thumbImages,
-                String deliveryInfo, String deliveryFee, String detailSection, int stock, String point) {
+    public Dish(String id, String alt, String deliveryType, String title, String description, String normalPrice,
+                String sellingPrice, String badge, Long categoryId, String deliveryInfo, String deliveryFee, int stock,
+                String point, Image image) {
         this.id = id;
-        this.mainImage = mainImage;
         this.alt = alt;
         this.deliveryType = deliveryType;
         this.title = title;
@@ -40,21 +40,15 @@ public class Dish {
         this.sellingPrice = sellingPrice;
         this.badge = badge;
         this.categoryId = categoryId;
-        this.topImage = topImage;
-        this.thumbImages = thumbImages;
         this.deliveryInfo = deliveryInfo;
         this.deliveryFee = deliveryFee;
-        this.detailSection = detailSection;
         this.stock = stock;
         this.point = point;
+        this.image = image;
     }
 
     public String getId() {
         return id;
-    }
-
-    public String getMainImage() {
-        return mainImage;
     }
 
     public String getAlt() {
@@ -89,24 +83,12 @@ public class Dish {
         return categoryId;
     }
 
-    public String getTopImage() {
-        return topImage;
-    }
-
-    public String getThumbImages() {
-        return thumbImages;
-    }
-
     public String getDeliveryInfo() {
         return deliveryInfo;
     }
 
     public String getDeliveryFee() {
         return deliveryFee;
-    }
-
-    public String getDetailSection() {
-        return detailSection;
     }
 
     public int getStock() {
@@ -117,19 +99,14 @@ public class Dish {
         return point;
     }
 
-    public boolean checkStock(int orderSize) {
-        return stock >= orderSize;
-    }
-
-    public void updateStock(int orderSize) {
-        this.stock -= orderSize;
+    public Image getImage() {
+        return image;
     }
 
     @Override
     public String toString() {
         return "Dish{" +
                 "id='" + id + '\'' +
-                ", mainImage='" + mainImage + '\'' +
                 ", alt='" + alt + '\'' +
                 ", deliveryType='" + deliveryType + '\'' +
                 ", title='" + title + '\'' +
@@ -138,13 +115,11 @@ public class Dish {
                 ", sellingPrice='" + sellingPrice + '\'' +
                 ", badge='" + badge + '\'' +
                 ", categoryId=" + categoryId +
-                ", topImage='" + topImage + '\'' +
-                ", thumbImages='" + thumbImages + '\'' +
                 ", deliveryInfo='" + deliveryInfo + '\'' +
                 ", deliveryFee='" + deliveryFee + '\'' +
-                ", detailSection='" + detailSection + '\'' +
                 ", stock=" + stock +
                 ", point='" + point + '\'' +
+                ", image=" + image +
                 '}';
     }
 }
