+++ BE/src/main/java/com/codesquad/sidedish/domain/Image.java
@@ -1,0 +1,25 @@
+package com.codesquad.sidedish.domain;
+
+public class Image {
+
+    private String mainImage;
+    private String topImage;
+    private String thumbImages;
+    private String detailSection;
+
+    public String getMainImage() {
+        return mainImage;
+    }
+
+    public String getTopImage() {
+        return topImage;
+    }
+
+    public String getThumbImages() {
+        return thumbImages;
+    }
+
+    public String getDetailSection() {
+        return detailSection;
+    }
+}
